Parsoid's setMwApi needs the end URI of the host
The modified URI is the default VirtualBox IP address we're using.
This URI must be the endpoint URI that the user navigates to,
because when Parsoid serves up URIs for images (and probably other
resources) it uses this URI. So otherwise it goes looking for
images at http://127.0.0.1/<whatever>, which is on the user's
computer...doesn't work.