Fix input to cmd_tee; remove script param inputs (breaking)
Script inputs to cmd_tee included "|| exit 1" on the end,
causing the "||" to be passed as the first param, "exit"
as the second param, and "1" as the third. This caused
variables in the scripts to be overwritten by these input
params. Removed exit 1. Will need another (better) method
of checking for success.

Breaking change: Also removed all param inputs to scripts,
since they were prone to causing issues.