Change title banners to function
Fixes issue with yums.sh banner saying that it's the install.sh
banner. This was due to the fact that the title was being printed
by a separate script, printTitle.sh, and the title bar taken from
the name of the script calling that script. However, when we
switched to doing "source yums.sh" instead of "bash yums.sh" it
caused install.sh to be that script-name.