elastic-build-index skip unavailable commands; fix settings
elastic-build-indes.sh calls scripts related to SMW and CirrusSearch,
but on the initial install (install.sh), when the "Demo Wiki" is
created, these extensions are not yet available. The Demo Wiki is
required to bootstrap the rest of the install, so when wikis are
created now import-wikis.sh will check for the existence of SMW and
CirrusSearch before attempting to use them.

Also, ExtensionSettings.php had a missing semicolon. I'm dumb.