Fix permissions for services accessing MediaWiki (#514)
Check for presence of HTTP_X_FORWARDED_FOR to determine if request is from
a real user or if it is a request from a service. If it's a service, then
set permissions to be able to view/edit all. Real users will have
HTTP_X_FORWARDED_FOR, service request (that don't go through the load
balancer) do not.