Move Extensions.php away from rsynced local-config
Move Extensions.php out of the directory that is a copy of
/opt/meza/config/local-public, because it was getting deleted when that
directory was being rsynced. Create a new directory in config/core called
"app-ansible" which will hold any Ansible-generated files the app servers
need that don't fit anywhere else. Long term, the current config/local_app
should move to config/core/app-local-public, so the name indicates that
it's a copy of local-public needed by app servers. That is outside the
scope of this change, though.