Use Extension:SimpleMathJax instead of Extension:Math (#605)
Extension:Math after MediaWiki 1.25 does not allow client-side rendering
of Math via the MathJax javascript library. Without this it is required to
install a server side renderer such as Mathoid, texvc, or LaTeXML. All of
these are complicated. At some point down the road perhaps we'll add one
(or more) of these, but for now it is much simpler to maintain previous
functionality with this other extension.

Note that the original SimpleMathJax had two issues:

  1. The MathJax library was loaded from a CDN, which is a
     security/privacy concern
  2. MathJax JS and CSS was loaded with every page load, regardless of
     whether or not <math> or <chem> tags were used

As such, the GitHub jmnote/SimpleMathJax repo was forked to
jamesmontalvo3/SimpleMathJax. The changes made in this fork have been
pull-requested back into the trunk, but at this time have not been
accepted. If they are accepted it will have to be determined whether or
not we use an unknown repository or continue to maintain this one. If we
continue to maintain it, jamesmontalvo3 will transfer the repo to
enterprisemediawiki.