Add note to my.cnf.j2
I don't know if it's worth coding this 'feature', so I thought to add a line of comment into the template. Monolith environments with only 1 GB of memory can fail because MySQL takes up a lot of it before Parsoid gets installed.  If the performance schema is disabled, it shouldn't affect an intentionally low power "demo" or "development" environment, but will allow Meza deploy to complete.  Feel free to decline and tell me that this should be an 'issue' instead.