Adds new config m_mode for allowing development and production variance
In production, PHP files are never checked for modification on the filesystem.
They are simply read into the PHP Opcache, and used from cache. (fast) Also,
There is less verbose logging in production.
This switch allows for easier enabling of features like profiling that
may only ever be used in development.