Use 'remote_src'
When the parsoid host is not the controller, the src file is on the remote. Therefore, use remote_src: yes so that Ansible finds the file to copy.
Otherwise, this fails when the Parsoid host in inventory is not the same host as the controller.  In my example setup, I have two application servers and both are designated as Parsoid servers. One is the controller, the other app server is just an app server.