Don't recreate meza-ansible if user already exists (Revert #965)
PR #965 did some work to fix issues with restrictive umask settings. It
also stated that it "made getmeza.sh more idempotent" but one of the
changes was to make it so each run of the script attempted to create
user meza-ansible, and in the process recreated SSH keys for the user.
This is the opposite of idempotent, and must be undone. This is
currently breaking docker builds. #965 does not indicate if there was a
specific reason for making this change.