Move ansible/python38 excludes from meza to dnf/yum repos
These changes move the dnf/yum exlusions from within meza to the dnf/yum
repositories. The rationale for the changes is that placing exclusions in
the proper repositories prevents the meza base ansible task that updates all
packages on the system from updating ansible and python to versions provided
in EPEL and Rocky/Red Hat AppStream which could break meza.